fix: remove unnecessary type conversion in diskusage_unix.go
- Remove int64() conversion from stat.Bsize as it's already the correct type
- Keep necessary conversions for stat.Blocks and stat.Bavail with gosec ignore comments
- This resolves the final unconvert linting issue in CI pipeline

+++ internal/services/diskusage_unix.go
@@ -14,11 +14,11 @@ func getDiskUsageForPath(path string) (*DiskUsage, error) {
 	}
 
 	// Calculate sizes in bytes
-	blockSize := int64(stat.Bsize)
+	blockSize := stat.Bsize
 	//nolint:gosec // Safe conversion for disk space calculation
-	totalSize := int64(stat.Blocks) * blockSize
+	totalSize := int64(stat.Blocks) * int64(blockSize)
 	//nolint:gosec // Safe conversion for disk space calculation  
-	freeSize := int64(stat.Bavail) * blockSize
+	freeSize := int64(stat.Bavail) * int64(blockSize)
 
 	return &DiskUsage{
 		Free:  freeSize,
